VP-1677:  EditUserModal.vue is separated
